Clean up VoiceCommands file
Create commands folder for all command related files, create Utils
folder for utility related functionality like process running
