+++ src/ProjektDjAladar/Commands/Audio.cs
@@ -1,0 +1,11 @@
+using DSharpPlus.Lavalink;
+
+namespace ProjektDjAladar
+{
+    struct Audio
+    {
+        public LavalinkExtension Lava;
+        public LavalinkNodeConnection Node;
+        public LavalinkGuildConnection Conn;
+    }
+}
