+++ src/ProjektDjAladar/Commands/TrackRequest.cs
@@ -1,10 +1,5 @@
 ﻿using DSharpPlus.CommandsNext;
 using DSharpPlus.Lavalink;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ProjektDjAladar
 {
