+++ src/ProjektDjAladar/Utils/JsonSettings.cs
